Revert "为风车添加产生RF能量和输出RF能量的功能"
This reverts commit 2c2ca22bb74877e9cac4a771d2c329aafd3750f4.

+++ src/main/java/com/github/mrstop/stdemo/inventory/ContainerWindmill.java
@@ -29,17 +29,17 @@ public class ContainerWindmill extends Container {
     @Override
     public void onCraftGuiOpened(ICrafting iCrafting) {
         super.onCraftGuiOpened(iCrafting);
-        iCrafting.sendProgressBarUpdate(this, 0, this.tileEntityWindmill.getCurrentPower());
+        iCrafting.sendProgressBarUpdate(this, 0, (int) this.tileEntityWindmill.getCurrentPower());
     }
 
     @Override
     public void detectAndSendChanges() {
         super.detectAndSendChanges();
-        int tempCurrentPower = this.tileEntityWindmill.getCurrentPower();
+        float tempCurrentPower = this.tileEntityWindmill.getCurrentPower();
         for (Object crafter : this.crafters) {
             ICrafting iCrafting = (ICrafting) crafter;
             if (this.currentPower != tempCurrentPower) {
-                iCrafting.sendProgressBarUpdate(this, 0, tempCurrentPower);
+                iCrafting.sendProgressBarUpdate(this, 0, (int) tempCurrentPower);
             }
         }
         this.currentPower = tempCurrentPower;
